refactor: update participant type imports to simplify paths and reduce type issues
